Add files via upload
This fixes altitude it seems. Now all dots start at center, and it fills up left to right. Then once the log reaches the full length, the first item drops and the current altitude is added to the end. avoids the dot class error, missing file error, and loads altitude graph after reset. Saving this as a step, in case cleaning the code breaks it